pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-code-polisher"
version = "1.0.0"
description = "A multi-agent pipeline for code refactoring and documentation"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "langchain-groq",
    "python-dotenv",
    "pydantic",
    "GitPython",
    "langchain"
]

[project.scripts]
# This allows you to type 'polish' in the terminal to run the tool
polish = "polisher.main:run_orchestrator"